Reorganized files and a small fix for LowestHealthPercentBelowThreshold
