+++ src/heuristics/heuristic.py
@@ -1,0 +1,8 @@
+from src.debug.logger import Logger
+
+
+class Heuristic:
+    logger = Logger()
+
+    def __init__(self):
+        pass
